[fix] password reset interface 분리
